break circular references by splitting some files into two
