+++ package.json
@@ -50,6 +50,7 @@
     "ts-simple-ast": "17.1.1",
     "tslint": "5.11.0",
     "tslint-config-standard": "8.0.1",
+    "tslint-no-circular-imports": "^0.6.1",
     "typescript": "^3.3.3"
   },
   "tags": [
