+++ src/Array.ts
@@ -1,7 +1,7 @@
 import { Alternative1 } from './Alternative'
 import { Applicative, Applicative1, Applicative2, Applicative2C, Applicative3, Applicative3C } from './Applicative'
-import { Compactable1, Separated } from './Compactable'
-import { Either } from './Either'
+import { Compactable1, Separated } from './Compactable_'
+import { Either } from './Either_'
 import { Extend1 } from './Extend'
 import { FilterableWithIndex1 } from './FilterableWithIndex'
 import { Foldable2v1 } from './Foldable2v'
@@ -11,7 +11,7 @@ import { FunctorWithIndex1 } from './FunctorWithIndex'
 import { HKT, Type, Type2, Type3, URIS, URIS2, URIS3 } from './HKT'
 import { Monad1 } from './Monad'
 import { Monoid } from './Monoid'
-import { none, Option, some } from './Option'
+import { none, Option, some } from './Option_'
 import { getSemigroup, Ord, ordNumber, fromCompare } from './Ord'
 import { Plus1 } from './Plus'
 import { getArraySetoid, Setoid } from './Setoid'
