+++ src/ChainRec.ts
@@ -1,5 +1,5 @@
 import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain3C } from './Chain'
-import { Either } from './Either'
+import { Either } from './Either_'
 import { HKT, Type, Type2, Type3, URIS, URIS2, URIS3 } from './HKT'
 
 /**
