+++ src/Choice.ts
@@ -1,4 +1,4 @@
-import { Either } from './Either'
+import { Either } from './Either_'
 import { HKT2, Type2, Type3, URIS2, URIS3, URIS4, Type4 } from './HKT'
 import { Profunctor, Profunctor2, Profunctor3, Profunctor4 } from './Profunctor'
 import { Category, Category2, Category3 } from './Category'
