+++ src/Identity.ts
@@ -2,7 +2,7 @@ import { Alt1 } from './Alt'
 import { Applicative } from './Applicative'
 import { ChainRec1, tailRec } from './ChainRec'
 import { Comonad1 } from './Comonad'
-import { Either } from './Either'
+import { Either } from './Either_'
 import { Foldable2v1 } from './Foldable2v'
 import { Lazy, toString } from './function'
 import { HKT } from './HKT'
