+++ src/Map.ts
@@ -1,23 +1,23 @@
 import { Applicative } from './Applicative'
-import { Compactable2, Separated } from './Compactable'
-import { Either } from './Either'
+import { Compactable2, Separated } from './Compactable_'
+import { Either } from './Either_'
 import { FilterableWithIndex2C } from './FilterableWithIndex'
 import { Foldable2v2C, Foldable2v3, Foldable2v2, Foldable2v1, Foldable2v } from './Foldable2v'
 import { FoldableWithIndex2C } from './FoldableWithIndex'
 import { Predicate, phantom, tuple } from './function'
 import { HKT, Type, Type2, Type3, URIS, URIS2, URIS3 } from './HKT'
 import { Monoid } from './Monoid'
-import { Option, none, some } from './Option'
+import { Option, none, some } from './Option_'
 import { Ord } from './Ord'
 import { Setoid, fromEquals } from './Setoid'
 import { TraversableWithIndex2C } from './TraversableWithIndex'
 import { Unfoldable, Unfoldable1 } from './Unfoldable'
-import { Semigroup } from './Semigroup'
+import { Semigroup } from './Semigroup_'
 import { Witherable2C } from './Witherable'
 import { FunctorWithIndex2C } from './FunctorWithIndex'
 import { Functor2 } from './Functor'
 import { Traversable2v2C } from './Traversable2v'
-import { Filterable2 } from './Filterable'
+import { Filterable2 } from './Filterable_'
 
 declare module './HKT' {
   interface URI2HKT2<L, A> {
