+++ src/NonEmptyArray.ts
@@ -18,7 +18,7 @@ import { compose, toString, Refinement, Predicate } from './function'
 import { HKT } from './HKT'
 import { Monad1 } from './Monad'
 import { Monoid } from './Monoid'
-import { none, Option, some } from './Option'
+import { none, Option, some } from './Option_'
 import { Ord } from './Ord'
 import { fold, getJoinSemigroup, getMeetSemigroup, Semigroup } from './Semigroup'
 import { Setoid, getArraySetoid, fromEquals } from './Setoid'
