+++ src/ReaderTaskEither.ts
@@ -1,14 +1,14 @@
 import { Alt3 } from './Alt'
 import { Bifunctor3 } from './Bifunctor'
-import { Either } from './Either'
+import { Either } from './Either_'
 import { constant, constIdentity, Predicate, Refinement } from './function'
-import { IO } from './IO'
+import { IO } from './IO_'
 import { IOEither } from './IOEither'
 import { Monad3 } from './Monad'
 import { MonadIO3 } from './MonadIO'
 import { Reader } from './Reader'
 import * as readerT from './ReaderT'
-import { Task } from './Task'
+import { Task } from './Task_'
 import * as taskEither from './TaskEither'
 import TaskEither = taskEither.TaskEither
 import { MonadTask3 } from './MonadTask'
