+++ src/StrMap.ts
@@ -1,6 +1,6 @@
 import { Applicative, Applicative1, Applicative2, Applicative3 } from './Applicative'
-import { Compactable1, Separated } from './Compactable'
-import { Either } from './Either'
+import { Compactable1, Separated } from './Compactable_'
+import { Either } from './Either_'
 import { FilterableWithIndex1 } from './FilterableWithIndex'
 import { Foldable, Foldable1, Foldable2, Foldable3 } from './Foldable'
 import { Foldable2v1 } from './Foldable2v'
@@ -9,9 +9,9 @@ import { Predicate, Refinement, tuple } from './function'
 import { FunctorWithIndex1 } from './FunctorWithIndex'
 import { HKT, Type, Type2, Type3, URIS, URIS2, URIS3 } from './HKT'
 import { Monoid } from './Monoid'
-import { Option } from './Option'
+import { Option } from './Option_'
 import * as R from './Record'
-import { getLastSemigroup, Semigroup } from './Semigroup'
+import { getLastSemigroup, Semigroup } from './Semigroup_'
 import { Setoid, fromEquals } from './Setoid'
 import { TraversableWithIndex1 } from './TraversableWithIndex'
 import { Unfoldable, Unfoldable1 } from './Unfoldable'
