+++ src/These.ts
@@ -1,14 +1,14 @@
 import { Applicative } from './Applicative'
 import { Bifunctor2 } from './Bifunctor'
-import { Either } from './Either'
+import { Either } from './Either_'
 import { Foldable2v2 } from './Foldable2v'
 import { phantom, toString } from './function'
 import { Functor2 } from './Functor'
 import { HKT } from './HKT'
 import { Monad2C } from './Monad'
 import { Monoid } from './Monoid'
-import { none, Option, some } from './Option'
-import { Semigroup } from './Semigroup'
+import { none, Option, some } from './Option_'
+import { Semigroup } from './Semigroup_'
 import { Setoid, fromEquals } from './Setoid'
 import { Traversable2v2 } from './Traversable2v'
