+++ src/Tuple.ts
@@ -4,14 +4,14 @@ import { Bifunctor2 } from './Bifunctor'
 import { Chain2C } from './Chain'
 import { ChainRec2C } from './ChainRec'
 import { Comonad2 } from './Comonad'
-import { Either } from './Either'
+import { Either } from './Either_'
 import { Foldable2v2 } from './Foldable2v'
 import { phantom, toString } from './function'
 import { HKT } from './HKT'
 import { Monad2C } from './Monad'
 import { Monoid } from './Monoid'
 import { contramap as contramapOrd, getSemigroup as getOrdSemigroup, Ord } from './Ord'
-import { Semigroup } from './Semigroup'
+import { Semigroup } from './Semigroup_'
 import { Semigroupoid2 } from './Semigroupoid'
 import { Setoid, fromEquals } from './Setoid'
 import { Traversable2v2 } from './Traversable2v'
