+++ src/Validation.ts
@@ -1,17 +1,17 @@
 import { Alt2C } from './Alt'
 import { Applicative, Applicative2C } from './Applicative'
 import { Bifunctor2 } from './Bifunctor'
-import { Compactable2C, Separated } from './Compactable'
-import { Either } from './Either'
-import { Filterable2C } from './Filterable'
+import { Compactable2C, Separated } from './Compactable_'
+import { Either } from './Either_'
+import { Filterable2C } from './Filterable_'
 import { Foldable2v2 } from './Foldable2v'
 import { phantom, Predicate, toString, Refinement } from './function'
 import { Functor2 } from './Functor'
 import { HKT } from './HKT'
 import { Monad2C } from './Monad'
 import { Monoid } from './Monoid'
-import { Option } from './Option'
-import { Semigroup } from './Semigroup'
+import { Option } from './Option_'
+import { Semigroup } from './Semigroup_'
 import { Setoid, fromEquals } from './Setoid'
 import { Traversable2v2 } from './Traversable2v'
 import { Witherable2C } from './Witherable'
